requirements.txt
board==1.0
busio==0.7.0
adafruit_lis2mdl==3.0.0
adafruit_lsm303_accel==3.0.0
lirc
adafruit_pcf8591==1.0.19
RPi.GPIO==0.7.0
audioplayer==0.6
mutagen==1.47.0
playsound==1.3.0
